+++ test-doo/promenade/runner.cljs
@@ -10,13 +10,15 @@
 (ns promenade.runner
   (:require
     [doo.runner :refer-macros [doo-tests]]
-    [promenade.core-test]))
+    [promenade.core-test]
+    [promenade.util-test]))
 
 
 (enable-console-print!)
 
 (try
   (doo-tests
-    'promenade.core-test)
+    'promenade.core-test
+    'promenade.util-test)
   (catch js/Error e
     (.log js/Console (.-stack e))))